Swallow undeliverable RxJava exceptions
AntennaPod threads might throw NPEs after disposing because we set controllers to null.
