+++ core/src/free/java/de/danoeh/antennapod/core/ClientConfig.java
@@ -9,7 +9,6 @@ import de.danoeh.antennapod.core.preferences.UserPreferences;
 import de.danoeh.antennapod.core.service.download.AntennapodHttpClient;
 import de.danoeh.antennapod.core.storage.PodDBAdapter;
 import de.danoeh.antennapod.core.util.NetworkUtils;
-import de.danoeh.antennapod.core.util.exception.RxJavaErrorHandlerSetup;
 import de.danoeh.antennapod.core.util.gui.NotificationUtils;
 
 import java.io.File;
@@ -49,7 +48,6 @@ public class ClientConfig {
         NetworkUtils.init(context);
         AntennapodHttpClient.setCacheDirectory(new File(context.getCacheDir(), "okhttp"));
         SleepTimerPreferences.init(context);
-        RxJavaErrorHandlerSetup.setupRxJavaErrorHandler();
         NotificationUtils.createChannels(context);
         initialized = true;
     }
